% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regustab.R
\docType{package}
\name{stabplot-package}
\alias{stabplot}
\alias{stabplot-package}
\title{Stability Plots for Lasso Stability Selection}
\description{
Provides two diagnostic plots for assessing stability selection with Lasso.
The `Regustab` function plots stability values across the regularisation
parameter grid, supporting regularisation tuning based on selection stability.
It highlights and returns `lambda.min`, `lambda.1se`, and `lambda.stable`
when the maximum stability exceeds 0.75. Otherwise, it returns
`lambda.stable.1sd`, defined using a one-standard-deviation criterion.
The `Convstab` function plots stability values against the number of
subsamples to assess the convergence of the stability estimates and returns
variables with selection frequencies exceeding a specified threshold,
which defaults to 0.5.
}
\references{
Nouraie, M., & Muller, S. (2026). Stability-guided hyper-parameter tuning for stability selection. Communications in Statistics - Theory and Methods, 1–19.

Meinshausen, N., & Bühlmann, P. (2010). Stability selection. Journal of the Royal Statistical Society Series B: Statistical Methodology, 72(4), 417-473.

Nogueira, S., Sechidis, K., & Brown, G. (2018). On the stability of feature selection algorithms. Journal of Machine Learning Research, 18(174), 1-54.

\url{https://github.com/nogueirs/JMLR2018}

Tibshirani, R. (1996). Regression shrinkage and selection via the lasso. Journal of the Royal Statistical Society Series B: Statistical Methodology, 58(1), 267-288.
}
\seealso{
\link[=Regustab]{Regustab}, \link[=Convstab]{Convstab}
}
\author{
\strong{Maintainer}: Mahdi Nouraie \email{mahdinouraie20@gmail.com}

Authors:
\itemize{
  \item Samuel Muller
}

}
