% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convstab.R
\name{Convstab}
\alias{Convstab}
\title{Convstab}
\usage{
Convstab(x, y, B, alpha = 0.05, thr = 0.5)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A numeric vector of response values.}

\item{B}{An integer specifying the number of sub-samples.}

\item{alpha}{A numeric value specifying the level of significance.}

\item{thr}{A numeric value specifying the threshold for selection frequency.}
}
\value{
A plot displaying the stability values and corresponding confidence interval through sequential sub-sampling. `Convstab` also prints the variables selected with a selection frequency greater than the threshold value.
}
\description{
`Convstab` creates a plot displaying stability values along with confidence intervals, against the sequential sub-sampling index within stability selection. This plot aids in monitoring the convergence status of stability values.
The function uses `lambda.stable` to generate the plot; if `lambda.stable` is unavailable, it defaults to `lambda.stable.1sd`.
}
\examples{
\dontrun{
set.seed(123)
x <- matrix(rnorm(1000), ncol = 10)
# create beta based on the first 3 columns of x and some error
beta <- c(0.5, 0.4, 0.3, rep(0, 7))
y <- x \%*\% beta + rnorm(100)
B <- 200
Convstab(x, y, B)  # Example usage of the Convstab function
#output
#Variable Selection_Frequency
#1       x1               0.970
#2       x2               0.895

}
}
\references{
Meinshausen, N., & Bühlmann, P. (2010). Stability selection. Journal of the Royal Statistical Society Series B: Statistical Methodology, 72(4), 417-473.

Nogueira, S., Sechidis, K., & Brown, G. (2018). On the stability of feature selection algorithms. Journal of Machine Learning Research, 18(174), 1-54.

https://github.com/nogueirs/JMLR2018

Tibshirani, R. (1996). Regression shrinkage and selection via the lasso. Journal of the Royal Statistical Society Series B: Statistical Methodology, 58(1), 267-288.
}
\seealso{
\link[=stabplot]{stabplot}
}
