% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convstab.R
\name{convstab}
\alias{convstab}
\title{convstab}
\usage{
convstab(x, y, B, alpha = 0.05)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A numeric vector of response values.}

\item{B}{An integer specifying the number of sub-samples.}

\item{alpha}{A numeric value specifying the level of significance.}
}
\value{
A plot displaying the stability values and corresponding confidence interval through sequential sub-sampling.
}
\description{
This function produces a plot of stability values and corresponding confidence intercval through sequential sub-sampling of stability selection to facilitate monitoring the convergence status.
}
\examples{
\dontrun{
set.seed(123)
library(hdi)
data(riboflavin) # Load the riboflavin data
data <- as.data.frame(cbind(Y=riboflavin$y - 1, X=riboflavin$x)) # Convert the data to a data frame
rm(riboflavin) # Remove the original data to save memory
x <- as.matrix(data[, -1]) # Extract the predictors
y <- data[,1] # Extract the response
B = 100
convstab(x, y, B, alpha = 0.05)
}
}
\references{
Meinshausen, N., & Bühlmann, P. (2010). Stability selection. Journal of the Royal Statistical Society Series B: Statistical Methodology, 72(4), 417-473.

https://github.com/nogueirs/JMLR2018
}
