% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regustab.R
\name{Regustab}
\alias{Regustab}
\title{Regustab}
\usage{
Regustab(x, y, B)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A numeric vector of response values.}

\item{B}{An integer specifying the number of sub-samples.}
}
\value{
A plot displaying the relationship between lambda values and stability. `Regustab` also prints the values of `lambda.min`, `lambda.1se`, and `lambda.stable`. If `lambda.stable` is not available, the function will display `lambda.stable.1sd` instead.
}
\description{
This function creates a plot that displays stability values in relation to regularization values for LASSO through stability selection.
The plot highlights key lambda values, including `lambda.min`, `lambda.1se`, and `lambda.stable`. If `lambda.stable` is not available, the function will display `lambda.stable.1sd` instead. Regustab also prints the values of highlighted regularization values on the plot (`lambda.min`, `lambda.1se`, and `lambda.stable` or `lambda.stable.1sd`).
}
\examples{
\dontrun{
set.seed(123)
x <- matrix(rnorm(1000), ncol = 10)
# create beta based on the first 3 columns of x and some error
beta <- c(1, 2, 3, rep(0, 7))
y <- x \%*\% beta + rnorm(100)
B <- 10
Regustab(x, y, B)  # Example usage of the Regustab function
#output
$min
[1] 0.07609021
$`1se`
[1] 0.2550241
$stable
[1] 0.3371269

}

}
\references{
Meinshausen, N., & Bühlmann, P. (2010). Stability selection. Journal of the Royal Statistical Society Series B: Statistical Methodology, 72(4), 417-473.

Nogueira, S., Sechidis, K., & Brown, G. (2018). On the stability of feature selection algorithms. Journal of Machine Learning Research, 18(174), 1-54.

https://github.com/nogueirs/JMLR2018

Tibshirani, R. (1996). Regression shrinkage and selection via the lasso. Journal of the Royal Statistical Society Series B: Statistical Methodology, 58(1), 267-288.
}
\seealso{
\link[=stabplot]{stabplot}
}
