% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regustab.R
\name{Regustab}
\alias{Regustab}
\title{Regustab}
\usage{
Regustab(x, y, B)
}
\arguments{
\item{x}{A numeric matrix of predictors.}

\item{y}{A numeric vector of response values.}

\item{B}{An integer specifying the number of sub-samples.}
}
\value{
A plot displaying the relationship between lambda values and stability.
}
\description{
This function produces a plot of stability values against regularization values for LASSO through stability selection to facilitate regularization tuning.
}
\examples{
\dontrun{
set.seed(123)
x <- matrix(rnorm(1000), ncol = 10)
# create beta based on the first 3 columns of x and some error
beta <- c(1, 2, 3, rep(0, 7))
y <- x \%*\% beta + rnorm(100)
B <- 10
Regustab(x, y, B)  # Example usage of the regustab function
}

}
\references{
Meinshausen, N., & Bühlmann, P. (2010). Stability selection. Journal of the Royal Statistical Society Series B: Statistical Methodology, 72(4), 417-473.

https://github.com/nogueirs/JMLR2018
}
